+++ src/buildifier/configuration.ts
@@ -1,0 +1,12 @@
+
+/**
+ * Configuration for the buildifier module.
+ */
+export type BuildifierConfiguration = {
+    owner: string,
+    repo: string,
+    releaseTag: string,
+    executable: string,
+    fixOnFormat: boolean,
+    verbose: number,
+};