+++ src/test/unit/bazeldoc/hover.test.ts
@@ -4,31 +4,18 @@
  *--------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { makeDocEntryLink, makeBazelDocGroupHoverMarkdown } from '../../bazelDocGroupHover';
+import { makeBazelDocGroupHoverMarkdown } from '../../../bazeldoc/hover';
 
 suite('StarlarkDocGroup Tests', () => {
 
     test('makeDocEntryLink', () => {
-        const actual = makeDocEntryLink('bar', {
-            name: 'built-in foos',
-            path: 'be/foo.html',
-            items: ['bar'],
-        });
-        const expected = `[bar](https://docs.bazel.build/versions/master/be/foo.html#bar)`;
-
-		assert.equal(actual, expected);
-	});
-
-	test('makeBazelDocGroupHoverMarkdown', () => {
         const actual = makeBazelDocGroupHoverMarkdown('bar', {
             name: 'built-in foos',
             path: 'be/foo.html',
             items: ['bar', 'baz'],
-        }).value;
+        }, 'https://docs.bazel.build/versions/master').value;
         const expected = `
-[bar](https://docs.bazel.build/versions/master/be/foo.html#bar) a member of the group **built-in foos**
-
----
+**[bar](https://docs.bazel.build/versions/master/be/foo.html#bar)** is a member of _built-in foos_
 
 [bar](https://docs.bazel.build/versions/master/be/foo.html#bar), [baz](https://docs.bazel.build/versions/master/be/foo.html#baz)`;